Certora Signature Specs Updated

+++ certora/harnesses/SafeHarness.sol
@@ -16,18 +16,18 @@ contract SafeHarness is Safe {
     }
 
     // harnessed functions
-    function signatureSplitPublic(bytes memory signatures, uint256 pos) public pure returns (uint8 v, bytes32 r, bytes32 s) {
+    function signatureSplitPublic(bytes memory signatures, uint256 pos) public pure returns (uint256 v, bytes32 r, bytes32 s) {
         require(signatures.length >= 65 * (pos + 1));
         return signatureSplit(signatures, pos);
     }
 
-    function getCurrentOwner(bytes32 dataHash, uint8 v, bytes32 r, bytes32 s) public pure returns (address currentOwner) {
+    function getCurrentOwner(bytes32 dataHash, uint256 v, bytes32 r, bytes32 s) public pure returns (address currentOwner) {
         if (v == 0 || v == 1) {
             currentOwner = address(uint160(uint256(r)));
         } else if (v > 30) {
-            currentOwner = ecrecover(keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", dataHash)), v - 4, r, s);
+            currentOwner = ecrecover(keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", dataHash)), uint8(v - 4), r, s);
         } else {
-            currentOwner = ecrecover(dataHash, v, r, s);
+            currentOwner = ecrecover(dataHash, uint8(v), r, s);
         }
     }
